requirements.txt
matplotlib
seaborn
numpy==1.23.5
pandas==1.5.3
scipy==1.10.0
scikit-learn==1.3.1
google-cloud-bigquery
google-cloud-storage
db-dtypes
pyarrow
# Trick to install the version of Tensorflow depending on your processor: darwin == Mac, ARM == M1
tensorflow==2.15.0
# Lib for Vision Transformer (ViT)
tensorflow_addons
vit-keras
# prevent bq 3+ db-dtypes package error when running training from bq
mlflow==2.1.1
prefect==2.14.9
python-dotenv
psycopg2-binary
vit_keras
tensorflow_addons
opencv-python
scikit-image
streamlit
fastapi
uvicorn
